+++ src/main/java/de/tum/in/www1/jenkins/notifications/model/Failure.java
@@ -1,4 +1,4 @@
-package io.jenkins.plugins.sample.model;
+package de.tum.in.www1.jenkins.notifications.model;
 
 import org.kohsuke.stapler.export.Exported;
 import org.kohsuke.stapler.export.ExportedBean;
@@ -7,8 +7,8 @@ import javax.xml.bind.annotation.XmlAttribute;
 import javax.xml.bind.annotation.XmlRootElement;
 
 @ExportedBean(defaultVisibility = 4)
-@XmlRootElement(name = "error")
-public class Error {
+@XmlRootElement(name = "failure")
+public class Failure {
     @XmlAttribute
     private String message;
     @XmlAttribute
