+++ src/main/resources/de/tum/in/www1/jenkins/notifications/SendTestResultsNotificationPostBuildTask/config.properties
